Add courses to course list
Added the ability to update the course list and add courses. Also misc fixes along the way

+++ src/app/view-and-edit/view-and-edit.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { RequestService } from 'src/app/request.service';
 import { ListSchedules } from 'src/app/Models/list-schedules.model';
 import { BrowserModule } from '@angular/platform-browser'
+import { AngularFireAuth } from "@angular/fire/auth";
+import firebase from "firebase/app";
 
 @Component({
   selector: 'app-view-and-edit',
@@ -19,30 +21,39 @@ export class ViewAndEditComponent implements OnInit {
   }
 
   deleteAllSchedules(){
-    this.service.deleteAllSchedules().subscribe(e =>{
-      alert(e);
-      this.displaySchedules();
-    },error => {
-      alert(error.error)
-    })
-  }
-
-  deleteSchedule(){
-    this.service.deleteSchedule(this.deleteScheduleName).subscribe(e =>{
-      alert(e);
-      this.displaySchedules();
-    },error => {
-      alert(error.error)
-    })
-  }
-
-  displaySchedules() {
-    this.service.listSchedules().subscribe(e => {
-        //const entries = Object.entries(e);
-        //console.log("result: " + entries);
-        this.schedules = e;
+    if(confirm("Confirm deleting all course lists?")) {
+      this.service.deleteAllSchedules().subscribe(e =>{
+        alert(Object.values(e));
+        this.displaySchedules();
       },error => {
         alert(error.error)
       })
     }
+  }
+
+  deleteSchedule(){
+    if(confirm("Are you sure you would like to delete: " + this.deleteScheduleName)) {
+      this.service.deleteSchedule(this.deleteScheduleName).subscribe(e =>{
+        alert(Object.values(e));
+        this.displaySchedules();
+      },error => {
+        alert(error.error)
+      })
+    }
+  }
+
+  displaySchedules() {
+    this.service.listSchedules().subscribe(e => {
+      var user = firebase.auth().currentUser;
+      var email = user.email;
+      for (let [key, value] of Object.entries(e)) {
+        if(Object.entries(value)[0][1] !=  email){
+          delete e[key];
+        }
+      }
+        this.schedules = e;
+      },error => {
+        alert(error.error)
+      })
+  }
 }
